requirements.txt
Django==5.2
djangorestframework
djangorestframework-simplejwt
channels
channels-redis
psycopg2-binary
django-cors-headers
drf-spectacular
python-dotenv
PyJWT
daphne
